Move MatrixRTCMode enum from settings.ts to ConfigOptions.ts
